interfaces/index: Switch placement of title tag
It's recommended to put the primary keyword/phrase the page is about,
first, and the secondary keyword/phrase the site is about, second. In
this case, it would be 'Interfaces' (primary) and 'NanoSat MO Framework'
(secondary).

Note: 'NanoSat MO Framework' is the primary keyword/phrase for the root
homepage of the site, so the other benefit here is that by using it as
the secondary keyword on this page, it won't compete against the
homepage to rank higher, but will more likely rank as the first result
for 'Interfaces' associated with NMF when that search query is used on
Google or other search engines.